Ready to use version is ready!

+++ build.sh
@@ -2,6 +2,9 @@
 
 #set -x
 
+bold=`tput bold`
+normal=`tput sgr0`
+
 except()
 {
 	case "$1" in
@@ -16,14 +19,25 @@ except()
 	esac
 }
 
+in_stages=()
+
+check_in_array()
+{
+	for el in "${@:2}"
+		do
+			[[ "$1" == "${el}" ]] && echo "1" && return 0
+		done
+	echo "0"
+	return 1
+}
+
 main()
 {
 	pushd "package"
-	tarName=$(find . -name "*.tar.gz")
+	tarName=$(find . -regex ".*/.*[^\(orig\|debian\)].tar.gz")
 
-	if [[ "${tarName}" == "" ]]
-		then
-			exit 1
+	if [[ "${tarName}" == "" ]]; then
+		exit 1
 	fi
 
 	tarName=${tarName#./}
@@ -33,51 +47,66 @@ main()
 	packageName=${packageName//[_]/-}
 
 	if echo "$packageName" | grep -q "python-"; then
-		mkdir "${packageName}"
 		dirName="${packageName}"
 	else
- 		mkdir "python-${packageName}"
- 		dirName="python-${packageName}"
+		dirName="python-${packageName}"
 	fi
 
-	pushd "${dirName}"
-	tar -xzvf "../${tarName}"
-	
-	buf=${tarName%.tar.gz}
-	mv ${buf}/* .
-	rm -rf ${buf}
-	
-	dh_make -e ${email} -f "../${tarName}" -s -y
+	popd
 
-	#sleep 15
-	popd
-	popd
-	python builder.py -c "config.yaml" > builder.log
-	pushd "package"
-	pushd "${dirName}"
-	
-	set DEB_BUILD_OPTIONS=nocheck
-	dpkg-buildpackage -rfakeroot -us -uc
-	
-	popd
-	deb_tar_gz=$(find . -name "*.debian.tar.gz")
-	orig_tar_gz=$(find . -name "*.orig.tar.gz")
-	tar -xzvf ${deb_tar_gz}
-	tar -xzvf ${orig_tar_gz}
-	pushd "debian"
-	for i in $(find . -regextype posix-egrep -regex ".*(EX|ex|README.Debian)$"); do rm -f $i; done
-	popd
-	rm -rf "python-${packageName}"
-	for i in $(find . -regextype posix-egrep -regex ".*(dsc|changes|debian.tar.gz|orig.tar.gz)$"); do rm -f $i; done
+	if [[ $(find ./package/* -type d -name "*") == "" ]]; then
+		pushd "package"
+		mkdir "${dirName}"
+		pushd "${dirName}"
+		tar -xzvf "../${tarName}"
+		
+		buf=${tarName%.tar.gz}
+		mv ${buf}/* .
+		rm -rf ${buf}
+		popd
+		popd
+	fi
+
+	if [[ "$(check_in_array 1 ${in_stages[@]})" == "1" ]]; then
+		pushd "package"
+		pushd "${dirName}"
+
+		dh_make -e ${email} -f "../${tarName}" -s -y
+
+		popd
+		popd
+	fi
+
+	if [[ "$(check_in_array 2 ${in_stages[@]})" == "1" ]]; then
+		python builder.py -c "config.yaml" > builder.log
+	fi
+
+	if [[ "$(check_in_array 3 ${in_stages[@]})" == "1" ]]; then
+		pushd "package"
+		pushd "${dirName}"
+		
+		set DEB_BUILD_OPTIONS=nocheck
+		dpkg-buildpackage -rfakeroot -us -uc
+		
+		popd
+		deb_tar_gz=$(find . -name "*.debian.tar.gz")
+		orig_tar_gz=$(find . -name "*.orig.tar.gz")
+		tar -xzvf ${deb_tar_gz}
+		tar -xzvf ${orig_tar_gz}
+		pushd "debian"
+		for i in $(find . -regextype posix-egrep -regex ".*(EX|ex|README.Debian)$"); do rm -f $i; done
+		popd
+		rm -rf "${dirName}"
+		for i in $(find . -regextype posix-egrep -regex ".*(dsc|changes|debian.tar.gz|orig.tar.gz)$"); do rm -f $i; done
+	fi
 	#tar -cvf ${deb_tar_gz} "debian"
 	#rm -rf "debian"
-	echo "Done!"
+	echo "${bold}Done!${normal}"
 }
 
-if [ $# -eq 0 ]
-	then
-		echo "Please, specify arguments"
-		exit 1
+if [ $# -eq 0 ]; then
+	echo "Please, specify arguments"
+	exit 1
 fi
 
 while [ $# -gt 0 ]
@@ -85,26 +114,25 @@ do
 	case "$1" in
 		-e)
 			shift
-			if [ $# -gt 0 ]
-				then
-					email="$1"
-				else
-					except -e
+			if [ $# -gt 0 ]; then
+				email="$1"
+			else
+				except -e
 			fi
 			shift
 			;;
 		-s)
 			shift
-			if [ $# -gt 0 ]
-				then
-					email="$1"
-				else
-					except -s
+			if [ $# -gt 0 ]; then
+				in_stages+=("$@")
+			else
+				except -s
 			fi
+			;;
 		*)
 			shift
 			;;
 	esac
 done
 
-$(main)
+main
